@@ -5,28 +5,33 @@
         "build": "vite build"
     },
     "devDependencies": {
-        "@testing-library/jest-dom": "^5.16.2",
-        "@testing-library/react": "^11.2.7",
-        "@testing-library/user-event": "^12.8.3",
-        "@vitejs/plugin-react": "^2.0.1",
+        "@types/react": "^18.0.16",
+        "@types/react-dom": "^18.0.6",
+        "@typescript-eslint/eslint-plugin": "^5.32.0",
+        "@typescript-eslint/parser": "^5.32.0",
+        "@vitejs/plugin-react": "^2.0.0",
         "@vitejs/plugin-react-refresh": "^1.3.6",
+        "autoprefixer": "^10.4.8",
         "axios": "^0.27",
+        "eslint": "^8.21.0",
+        "eslint-config-prettier": "^8.5.0",
         "laravel-vite-plugin": "^0.5.0",
         "lodash": "^4.17.19",
-        "postcss": "^8.1.14",
-        "react": "^17.0.2",
-        "react-beautiful-dnd": "^13.1.0",
-        "react-dom": "^17.0.2",
-        "react-scripts": "5.0.0",
-        "vite": "^3.0.0",
-        "web-vitals": "^0.2.4"
+        "postcss": "^8.4.16",
+        "prettier": "^2.7.1",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-query": "^4.0.0",
+        "react-router-dom": "^5.3.0",
+        "tailwindcss": "^3.1.8",
+        "vite": "^3.0.0"
     },
     "dependencies": {
+        "@chakra-ui/react": "^2.2.8",
         "@emotion/react": "^11.10.0",
         "@emotion/styled": "^11.10.0",
-        "@material-ui/core": "^4.12.4",
-        "@mui/icons-material": "^5.8.4",
-        "@mui/material": "^5.10.1",
-        "react-router-dom": "^6.3.0"
+        "@types/react-router-dom": "^5.3.0",
+        "framer-motion": "^7.2.0",
+        "sass": "^1.54.3"
     }
 }
