@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\auth;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
