@@ -3,14 +3,14 @@ import {
   Link
 } from "react-router-dom";
 
-import { Button } from '@material-ui/core';
+import { Button } from '@chakra-ui/react'
 
 export const NotFound = () => {
   return (<>
     <h1>お探しのページは見つかりませんでした。</h1>
     <br />
     <Link to={`/`}>
-      <Button color="secondary" variant="contained">ホームに戻る</Button>
+      <Button colorScheme='red'>ホームに戻る</Button>
     </Link>
   </>);
 };
