@@ -1,7 +1,10 @@
 import React from "react";
 
+import { Heading } from '@chakra-ui/react'
+
+
 export const Home = () => {
   return (<>
-    <h2>Home</h2>
+    <Heading as='h1' size='md'>Home</Heading>
   </>);
 };
